refactor(web): move the re-auth predicate beside the provider table
`canReauthenticateWith` is a predicate over accounts, so the gate had to
import a "use client" component module to reach it. It now sits with
`isSocialProvider` in `@/lib/auth/social-providers`, and its tests moved
with it.

Disable Confirm until the session resolves: the email came from
`useSession`, so an early submit posted an empty one and reported a wrong
password for a correct one.

Collapse the three `setLoading(false)` branches in `handleDisconnect` into
one `finally`, stop exporting the session-not-fresh code that nothing
outside its own module reads, and drop the test fixtures the removed
resume mechanism needed.
