+++ apps/web/src/lib/auth/social-providers.tsx
@@ -1,3 +1,4 @@
+import type { Account } from "@sokosumi/utils";
 import type { ReactNode } from "react";
 
 import { GoogleIcon, MicrosoftIcon } from "@/components/social-icons";
@@ -20,4 +21,13 @@ export function isSocialProvider(
   providerId: string,
 ): providerId is SocialProvider {
   return SOCIAL_PROVIDERS.some((provider) => provider === providerId);
+}
+
+/** True when a viewer owns at least one method the re-auth dialog can offer. */
+export function canReauthenticateWith(accounts: Account[]): boolean {
+  return accounts.some(
+    (account) =>
+      account.providerId === AccountProvider.CREDENTIAL ||
+      isSocialProvider(account.providerId),
+  );
 }
